+++ frontend/components/cytoscape.vue
@@ -22,6 +22,7 @@ import { useCytoscapeStore } from "@store/useCytoscapeStore";
 import CONSTANTS from "@cons/constants";
 import Filter from "@comp/units/filter.vue";
 import SelectLayout from "@comp/units/selectLayout.vue";
+import useCytoscapeDragOpt from "../composables/cytoscape/dragOpt";
 
 const nuxtApp = useNuxtApp();
 
@@ -64,97 +65,8 @@ interface Filter {
   isUnChecked: boolean;
 }
 
-const options = ref({
-  grabbedNode: (node) => {
-    // 노드를 선택할 때 마다, 이벤트를 생성해주고 있기 때문에, 노드 선택시 이전 이벤트를 지우고 시작한다. (이벤트가 복수개로 설정되버림.)
-    node.off("cdndout");
-
-    node.on("cdndout", (event, parentNode, dropSibling) => {
-      const cy = event.cy;
-      const thisNode = event.target;
-
-      /**
-       * dragOut 할때 해야할 일
-       * 1. 선택한 노드가 부모 노드(이하 카테고리)에 포함되어 있는 경우, 해당 카테고리의 edges 정보를 물려받아야 함.
-       * 2. 선택한 노드의 카테고리에 선택한 노드를 제외한 노드가 1개 이하인 경우,
-       *    2-1. 다른 하위 노드가 있는 경우, 해당 노드에게 카테고리의 edges 정보를 물려줘야함.
-       *    2-2. 다른 하위 노드의 parent 정보를 삭제해 줘야함. (하위노드 move())
-       *    2-3. 분류를 삭제해 줘야함.
-       */
-
-      // step1. 부모 노드의 edges 정보를 저장해둔다.
-      let targets = parentNode
-        .connectedEdges()
-        .map((e) =>
-          parentNode.data().id === e.data().source ? e.data().target : ""
-        );
-
-      // edges 정보를 물려줄 node 정보를 array 에 저장해두는 변수 선언
-      // 선택한 노드는 dragout 하는 노드이기 때문에, children 에 넣어둔다.
-      let children = [thisNode];
-
-      // 현재 노드를 dragout 한 후에 남아있는 노드를 구한다.
-      // 선택한 노드는 이미 부모 노드에서 빠져있다.
-      if (parentNode.children().length < 2) {
-        // 부모노드에 남아있는 자식 노드가 1개 이하인 경우
-        // 부모노드의 존재 의미가 없기 때문에 부모노드를 삭제한다.
-        // 삭제하기 전에 자식노드를 독립
-        children = children.concat(parentNode.children());
-
-        // 부모노드의 자식노드에서 해당 parent 정보를 지워준다.
-        parentNode.children().move({ parent: null });
-      }
-
-      // 부모노드에 자식노드가 없다면, 부모노드를 삭제한다.
-      if (!parentNode.isParent()) {
-        parentNode.remove();
-      }
-
-      // children 변수에는 부모노드의 edges 정보를 물려받아야 하는 노드들이 들어있다. 반영해준다.
-      let newEdges = [];
-      children.forEach((e) => {
-        newEdges = newEdges.concat(
-          targets.map((t, ti) => {
-            return {
-              group: "edges",
-              data: {
-                id: `${e.data().id}_${t}_${ti}`,
-                source: e.data().id,
-                target: t
-              }
-            };
-          })
-        );
-      });
-      cy.add(newEdges);
-    });
-
-    return true;
-  },
-  dropTarget: (dropTarget, grabbedNode) => {
-    // grabbedNode : 내가 선택한 노드
-    // dropTarget : 내가 drop 할수 있는 모든 노드.
-    // ->  분류를 추가하는 기능은 지원하지 않음 === 어떤 노드에도 drop 할 수 없음. 위치 이동만 가능함.
-    return false;
-  },
-  dropSibling: (dropSibling, grabbedNode) => {
-    // grabbedNode : 내가 선택한 노드
-    // dropSibling : 고아노드를 제외한 내가 drop 할 수 있는 노드
-    // dropTarget 은 false 고 dropSibling 은 true 일 수 없음.
-    return false;
-  },
-  newParentNode: (grabbedNode, dropSibling) => {
-    // 고아노드 + 고아노드 를 합칠때, 부모노드가 생성될때 호출 됨.
-    // 노드 + 노드 를 설정할 수 없도록 했기 때문에 이 option method 에서는 아무것도 할 필요가 없음.
-    return {};
-  },
-  boundingBoxOptions: {
-    includeOverlays: false,
-    includeLabels: false
-  },
-  overThreshold: 200,
-  outThreshold: 200
-});
+const hideNodeList = ref([]);
+const parentColorObj = ref({});
 
 const mmContainer = ref(null);
 const filters = ref<Filter[]>([]);
@@ -173,9 +85,6 @@ onMounted(() => {
     layout: cytoscapeStore.graphLayouts[props.defaultLayout]
   });
 
-  // layout 처리 작업중
-  // cy.layout(fcoseLayoutOpt);
-
   cy.zoom({
     // level: 2.0,
     renderedPosition: {
@@ -183,8 +92,9 @@ onMounted(() => {
       y: container.offsetHeight / 2
     }
   });
-  const cdnd = cy.compoundDragAndDrop(options.value);
+  const cdnd = cy.compoundDragAndDrop(useCytoscapeDragOpt().option.value);
 
+  // 부모 노드 tag 추가
   cy.nodeHtmlLabel([
     {
       query: ":parent",
@@ -195,7 +105,9 @@ onMounted(() => {
       tpl: (data) => {
         const el = cy.getElementById(data.id);
 
-        return `<p>${
+        return `<p class="htmlTag ${
+          hideNodeList.value.includes(data.id) ? "hide" : ""
+        }" id="htmlTag_${data.id}">${
           el.isParent()
             ? `${el.data().label} (${el.children().length})`
             : el.data().label
@@ -203,16 +115,37 @@ onMounted(() => {
       }
     }
   ]);
+
   cy.elements().forEach((element: any) => {
     let colorCode: string = getRandomColor();
+    if (element.isChild()) {
+      // 자식노드는 별도 처리한다.
+      return;
+    }
+    if (element.isParent()) {
+      // 자식 노드 색상을 여기서 설정해둔다.
+      element.data("childColor", getRandomColor());
+    }
+
     if (element.isNode()) {
+      // 자식 노드 일 경우, 색상을 통일시켜줘야함.
       element.style("background-color", colorCode);
-      filters.value.push({
-        id: element.id(),
-        label: element.data().label,
-        color: colorCode,
-        isUnChecked: false
-      });
+      // 자식노드는 filter 에 표시하지 않음.
+      if (!element.isChild()) {
+        filters.value.push({
+          id: element.id(),
+          label: element.data().label,
+          color: colorCode,
+          isUnChecked: false
+        });
+      }
+    }
+  });
+
+  cy.elements().forEach((e) => {
+    // 자식노드일때, 부모노드에 설정해두었던 색상코드를 일괄적용한다.
+    if (e.isChild()) {
+      e.style("background-color", e.parent().data("childColor"));
     }
   });
 
@@ -248,7 +181,25 @@ onMounted(() => {
       } else if (element.isEdge()) {
         element.addClass(CONSTANTS.NOT_SELECTED);
       }
+
+      // loop 를 도는 노드가 자식노드이고,
+      // 해당 노드의 부모 노드가 선택한 노드라면,
+      // loop를 도는 노드도 선택한걸로 해줘야함.
+      if (
+        element.isChild() &&
+        element.parent().data().id === evtTarget.data().id
+      ) {
+        element.removeClass(CONSTANTS.NOT_SELECTED);
+        element.addClass(CONSTANTS.HIGHLIGHT);
+      }
     });
+
+    if (evtTarget.isChild()) {
+      // 자식노드면 부모 노드도 선택한걸로 처리해줘야함.
+      highlightedNodeId.value.push(evtTarget.parent().data().id);
+      evtTarget.parent().removeClass(CONSTANTS.NOT_SELECTED);
+      evtTarget.parent().addClass(CONSTANTS.HIGHLIGHT);
+    }
   });
 
   cy.on("tap", "edge", (event: any) => {
@@ -301,6 +252,15 @@ const getRandomColor = () => {
 };
 
 const setNodeView = (id: string, isUnChecked: boolean) => {
+  const htmlTagEl = document.querySelector("p#htmlTag_" + id + ".htmlTag");
+  if (htmlTagEl !== undefined) {
+    if (isUnChecked) {
+      hideNodeList.value.push(id);
+    } else {
+      hideNodeList.value = hideNodeList.value.filter((e) => e !== id);
+    }
+  }
+
   cy.elements().forEach((element: any) => {
     if (element.isNode()) {
       if (element.id() === id) {
@@ -337,5 +297,9 @@ button {
 }
 .part {
   border: 1px solid red;
+}
+
+p.htmlTag.hide {
+  display: none;
 }
 </style>
